Refactored AuthLayout and useUserAuthenticated hook for easier state management
